+++ .gitignore
@@ -7,3 +7,5 @@ target
 **/*.rs.bk
 
 .direnv
+
+result
