+++ Cargo.toml
@@ -1,10 +1,10 @@
 [workspace]
-members = ["kernel", "ksync", "riscv"]
+members = ["kernel"]
 resolver = "3"
 
 [workspace.package]
 license = "GPL"
 edition = "2024"
 
-[workspace.dependencies]
-riscv = { path = "./riscv", default-features = false }
+# [workspace.dependencies]
+# riscv = { path = "./riscv", default-features = false }
