--- kernel/src/arch/riscv/mmu/sv39/mod.rs
@@ -1,9 +1,0 @@
-mod page_table;
-mod page_table_entry;
-mod physical_address;
-mod virtual_address;
-
-pub use page_table::*;
-pub use page_table_entry::*;
-pub use physical_address::*;
-pub use virtual_address::*;
