--- kernel/src/arch/riscv/trap/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod trap;
-pub(crate) mod trap_frame;
-pub(crate) mod trap_handler;
