--- kernel/src/driver/mod.rs
@@ -1,1 +1,0 @@
-pub mod uart;
