--- kernel/src/mm/mappings.rs
@@ -1,10 +1,0 @@
-use crate::arch::mmu::{PhysicalAddress, VirtualAddress};
-
-pub const KERNEL_IMAGE_START_VA: VirtualAddress =
-    unsafe { VirtualAddress::from_raw_unchecked(0xffff_ffff_8000_0000) };
-
-pub const KERNEL_IMAGE_START_PA: PhysicalAddress =
-    unsafe { PhysicalAddress::from_raw_unchecked(0x8000_0000) };
-
-pub const KERNEL_DIRECT_MAPPING_BASE: VirtualAddress =
-    unsafe { VirtualAddress::from_raw_unchecked(0xffff_ffd6_0000_0000) };
