--- kernel/src/task/mod.rs
@@ -1,10 +1,0 @@
-mod process;
-mod process_table;
-mod schedule;
-
-pub use process::*;
-pub use process_table::*;
-pub use schedule::*;
-
-pub const TASK_PID_IDLE: usize = 0;
-pub const TASK_PID_REAPER: usize = 1;
