--- kernel/src/userspace/mod.rs
@@ -1,17 +1,0 @@
-pub mod shell;
-pub mod syscalls;
-
-#[inline(always)]
-pub fn write<T: AsRef<[u8]>>(buf: T) -> isize {
-    syscalls::write(buf.as_ref().as_ptr(), buf.as_ref().len())
-}
-
-#[unsafe(no_mangle)]
-pub extern "C" fn userspace_sleep_print_loop() -> ! {
-    unsafe { core::arch::asm!(".align 12") };
-    loop {
-        syscalls::sleep_ms(1000);
-
-        let _ = write("[2] writing babeee");
-    }
-}
