--- kernel/src/userspace/syscalls.rs
@@ -1,72 +1,0 @@
-use core::arch::asm;
-
-use crate::syscall::{
-    SYSCALL_EXIT, SYSCALL_READ, SYSCALL_SHUTDOWN, SYSCALL_SLEEP_MS, SYSCALL_WRITE,
-};
-
-pub fn write(data_ptr: *const u8, len: usize) -> isize {
-    let ret: isize;
-    unsafe {
-        asm!(
-            "li a0, 1",
-            "ecall",
-            in("a7") SYSCALL_WRITE,
-            in("a1") data_ptr,
-            in("a2") len,
-            lateout("a0") ret,
-            options(nostack),
-        )
-    }
-
-    ret
-}
-
-pub fn read(buf: *mut u8, count: usize) -> isize {
-    let ret: isize;
-    unsafe {
-        asm!(
-            "li a0, 0",
-            "ecall",
-            in("a7") SYSCALL_READ,
-            in("a1") buf,
-            in("a2") count,
-            lateout("a0") ret,
-            options(nostack),
-        )
-    }
-
-    ret
-}
-
-pub fn sleep_ms(ms: usize) {
-    unsafe {
-        asm!(
-            "ecall",
-            in("a7") SYSCALL_SLEEP_MS,
-            in("a0") ms,
-            options(nostack)
-        )
-    }
-}
-
-// TODO: temporary syscall
-pub fn shutdown() {
-    unsafe {
-        asm!(
-            "ecall",
-            in("a7") SYSCALL_SHUTDOWN,
-            options(nostack)
-        )
-    }
-}
-
-pub fn exit(exit_code: i32) {
-    unsafe {
-        asm!(
-            "ecall",
-            in("a7") SYSCALL_EXIT,
-            in("a0") exit_code,
-            options(nostack)
-        )
-    }
-}
