--- ksync/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "ksync"
-version = "0.1.0"
-license.workspace = true
-edition.workspace = true
-
-[dependencies]
-
-[features]
-default = ["riscv"]
-riscv = []
