--- ksync/src/arch/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "riscv")]
-pub mod riscv;
-
-#[cfg(feature = "riscv")]
-pub use riscv::*;
