--- riscv/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "riscv"
-version = "0.1.0"
-license.workspace = true
-edition.workspace = true
-
-[dependencies]
-bitfields = "1.0.2"
